+++ library/book/views.py
@@ -1,8 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect, HttpResponseRedirect
 from django.contrib.auth.decorators import login_required
-from .models import author as auth, genre, publisher, book, stock
-from location.models import library_location
-from managementUser.models import admin_library as admin_library_model
+from .models import author as auth, genre, publisher, book
+from managementUser.models import admin_library
 from .form import form_create_author, form_edit_author, form_create_genre, form_edit_genre
 from .form import form_create_publisher, form_edit_publisher, form_create_book, form_edit_book
 from .form import form_create_stock
@@ -47,10 +46,12 @@ def author_update(response, id):
             instance = form.save(commit=False)  # Prevent premature saving
 
             # Handle empty values explicitly:
+            '''
             if not instance.middle_name:
                 instance.middle_name = ''  # Set to None for empty strings
             if not instance.last_name:
                 instance.last_name = '' 
+            '''
             instance.save()
             return redirect('author_read')
     
@@ -197,31 +198,23 @@ def book_create(response):
         return HttpResponseRedirect('/')
     
     if response.method == "POST":
-        form = form_create_book(response.POST)
-        form2 = form_create_stock(user=response.user, data=response.POST)
-
+        form = form_create_book(data = response.POST, user=response.user)
         if form.is_valid():
             instance = form.save(commit=False)  # Prevent premature saving
-            instance.save()
 
-            if form2.is_valid():
-                stock_instance = form2.save(commit=False) # Prevent premature saving
-                stock_instance.book = instance
+            # Prevent unauthorization library admins
+            if response.user.role == 'admin':
+                admin_lib = admin_library.objects.filter(user=response.user).first()
+                if admin_lib.library_location != instance.library_location:
+                    return redirect('book_read')
                 
-                if not stock_instance.decsription:
-                    stock_instance.decsription = ''  # Set to None for empty strings
-
-                stock_instance.save()
-
-            else:
-                return redirect('book_read')
+            instance.save()
             
             return redirect('book_read')
         
     else:
-        form = form_create_book()
-        form2 = form_create_stock(user=response.user)
-    return render(response, "book/book_create.html", {"form":form, "form2":form2})
+        form = form_create_book(user=response.user)
+    return render(response, "book/book_create.html", {"form":form})
 
 @login_required()
 def book_update(response, id):
@@ -232,7 +225,12 @@ def book_update(response, id):
     if response.user.role not in author:
         return HttpResponseRedirect('/')
     
-    form = form_edit_book(response.POST or None, instance=data)
+    if response.user.role == 'admin':
+        admin_lib = admin_library.objects.filter(user=response.user).first()
+        if admin_lib.library_location != data.library_location:
+            return redirect('book_read')
+    
+    form = form_edit_book(data = response.POST or None, instance=data , user=response.user)
     if response.method =="POST":
         if form.is_valid():
             instance = form.save(commit=False)  # Prevent premature saving